airflow.requirements.txt
pandas
numpy
psycopg2-binary
apache-airflow-providers-amazon
pyarrow
vnstock3
minio
trino